requirements.txt
pyfiglet
speedtest-cli
